examples/llm-data-exfiltration-redteam-python/requirements.txt
tszclient-py>=0.1.0
python-dotenv